refactor: :art: Collected project components into one folder
